Remove unused utility scripts and reorganize messaging logic.
This commit removes redundant and outdated utility scripts related to messaging, request handling, and state management. These files are no longer needed after recent refactoring, simplifying the codebase for easier maintenance and improved clarity.
