+++ common/messaging/handlers/support_handler.py
@@ -4,7 +4,7 @@ import logging
 from typing import Dict, Any, Optional, Union
 
 from common.db.models import get_db_user_id_by_telegram_id
-from common.messaging.utils import safe_send_message, safe_send_menu
+from common.messaging.unified_platform_utils import safe_send_message, safe_send_menu
 
 logger = logging.getLogger(__name__)
