+++ services/scraper_service/app/tasks.py
@@ -12,11 +12,11 @@ from contextlib import contextmanager
 
 from common.db.database import execute_query
 from common.utils.s3_utils import _upload_image_to_s3
-from common.utils.req_utils import fetch_ads_flatfy
+from common.utils.unified_request_utils import fetch_ads_flatfy
 from common.config import GEO_ID_MAPPING_FOR_INITIAL_RUN, AWS_CONFIG, REDIS_URL
 from common.db.models import store_ad_phones
 from common.celery_app import celery_app
-from common.utils.request_utils import make_request
+from common.utils.unified_request_utils import make_request
 from common.utils.ad_utils import process_and_insert_ad
 
 # ---------------------------
